feat(xtoken): unify teacher-logits IPC producer for full + top-k
Same-vocab teachers previously shipped top-k logits via the driver path (get_topk_logits), which gathers dense [B, T, k] value/index tensors to the driver and re-shards them to the student every step (~16-19 GB/step for k=8192), inflating policy_training/submit_training_futures. Cross-tokenizer teachers already used zero-copy CUDA IPC.

Add a single producer, get_teacher_logits_ipc(mode='full'|'topk'), that handles both transports over persistent CUDA IPC buffers (1 buffer for full logits; 2 for top-k values+indices), generalizing the existing single-buffer machinery. Top-k values ship in native compute dtype (bf16) and indices in int32 (lossless for <2^31 vocab; cast to int64 at the gather). get_full_logits_ipc now delegates to mode='full'; same-vocab top-k teachers ship a teacher_{i}_topk_ipc handle list instead of dense tensors. Numerically bf16-exact vs the prior driver path.

+++ nemo_rl/algorithms/loss/loss_functions.py
@@ -1569,9 +1569,7 @@ class CrossTokenizerDistillationLossFn(LossFunction):
             )
             loss = kl_scale * total_kd + ce_loss
         else:
-            kl_scale = torch.tensor(
-                1.0, device=total_kd.device, dtype=total_kd.dtype
-            )
+            kl_scale = torch.tensor(1.0, device=total_kd.device, dtype=total_kd.dtype)
             loss = self.kl_loss_weight * total_kd + self.ce_loss_scale * ce_loss
 
         # Student next-token accuracy (quick per-step signal), masked to valid
@@ -1653,9 +1651,7 @@ class CrossTokenizerDistillationLossFn(LossFunction):
         prefix = f"alignment_{i}_"
         gold, xtoken = self._resolve_gold_xtoken(i, use_per_teacher_flags)
         if xtoken and not gold:
-            raise ValueError(
-                f"teacher {i}: xtoken_loss=True requires gold_loss=True."
-            )
+            raise ValueError(f"teacher {i}: xtoken_loss=True requires gold_loss=True.")
 
         if gold:
             kd, kl_common, l1_uncommon, num_valid_chunks, top1 = self._compute_gold(
@@ -1773,7 +1769,9 @@ class CrossTokenizerDistillationLossFn(LossFunction):
         T = self.temperature
         student_log_probs = torch.log_softmax(logits.float() / T, dim=-1)
         teacher_log_probs = torch.log_softmax(topk_logits.float() / T, dim=-1)
-        student_gathered = torch.gather(student_log_probs, -1, topk_indices)
+        # Indices arrive int32 over IPC (or int64 on the legacy path); gather
+        # requires int64.
+        student_gathered = torch.gather(student_log_probs, -1, topk_indices.long())
         student_log_probs_k = student_gathered - torch.logsumexp(
             student_gathered, dim=-1, keepdim=True
         )
@@ -1842,9 +1840,7 @@ class CrossTokenizerDistillationLossFn(LossFunction):
             if self.normalize_teacher_by_vocab:
                 v_scale = (
                     torch.log(
-                        torch.tensor(
-                            float(self.teacher_vocab_sizes[i]), device=device
-                        )
+                        torch.tensor(float(self.teacher_vocab_sizes[i]), device=device)
                     )
                     / temp_weight
                 )
